# apps/web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Smart Classroom Management API",
    "description": "Multi-source classroom occupancy, sensor fusion, and resource management engine.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/sensor-data": {
      "post": {
        "tags": [
          "ingest"
        ],
        "summary": "Ingest Sensor Data",
        "description": "Ingests raw sensor telemetry from either the sensor-sim or physical ESP32 (\u00a77.1).\nProtected by X-Device-Key header.",
        "operationId": "ingest_sensor_data_api_sensor_data_post",
        "parameters": [
          {
            "name": "X-Device-Key",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "title": "X-Device-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SensorEventSchema"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/classrooms/{classroom_id}/commands/pending": {
      "get": {
        "tags": [
          "ingest"
        ],
        "summary": "Get Pending Commands",
        "description": "Returns queued appliance commands awaiting execution by device (\u00a77.1).",
        "operationId": "get_pending_commands_api_classrooms__classroom_id__commands_pending_get",
        "parameters": [
          {
            "name": "classroom_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Classroom Id"
            }
          },
          {
            "name": "X-Device-Key",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "title": "X-Device-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PendingCommandResponse"
                  },
                  "title": "Response Get Pending Commands Api Classrooms  Classroom Id  Commands Pending Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/commands/{command_id}/ack": {
      "post": {
        "tags": [
          "ingest"
        ],
        "summary": "Acknowledge Command",
        "description": "Acknowledges appliance command execution by device (\u00a77.1).",
        "operationId": "acknowledge_command_api_commands__command_id__ack_post",
        "parameters": [
          {
            "name": "command_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Command Id"
            }
          },
          {
            "name": "X-Device-Key",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "title": "X-Device-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AckResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AckResponse": {
        "properties": {
          "acked": {
            "type": "boolean",
            "title": "Acked",
            "default": true
          }
        },
        "type": "object",
        "title": "AckResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "PendingCommandResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "classroom_id": {
            "type": "string",
            "title": "Classroom Id"
          },
          "device": {
            "type": "string",
            "title": "Device"
          },
          "command": {
            "type": "string",
            "title": "Command"
          }
        },
        "type": "object",
        "required": [
          "id",
          "classroom_id",
          "device",
          "command"
        ],
        "title": "PendingCommandResponse"
      },
      "SensorEventSchema": {
        "properties": {
          "classroom_id": {
            "type": "string",
            "title": "Classroom Id"
          },
          "device_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Device Id"
          },
          "pir_motion": {
            "type": "boolean",
            "title": "Pir Motion",
            "default": false
          },
          "event": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Event"
          },
          "temperature": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 60.0,
                "minimum": -10.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Temperature"
          },
          "humidity": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 100.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Humidity"
          },
          "ac_status": {
            "type": "boolean",
            "title": "Ac Status",
            "default": false
          },
          "light_status": {
            "type": "boolean",
            "title": "Light Status",
            "default": false
          },
          "wifi_rssi": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Wifi Rssi"
          },
          "source": {
            "type": "string",
            "enum": [
              "simulation",
              "esp32"
            ],
            "title": "Source",
            "default": "simulation"
          },
          "timestamp": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Timestamp"
          }
        },
        "type": "object",
        "required": [
          "classroom_id"
        ],
        "title": "SensorEventSchema"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}